added simpe_bed_merge.py to chain together adjacent bed regions with the same call or determined by identical 4th columns.
This is version used for Final_xo_chain_of_custody.md and it was used to call all regions between A and B regions. There are some ambig regions included as well that are not junctions per say. They occur at the beginning or the end of the scaffold. These are still worth checking on the map.